Adding Azure static web app deployment step
For a demo showcasing multiple deployment steps and environments, with as few moving
parts as possible, Azure static web apps is listed as the dev environment target. The
official action allows for the uploading of assets from a GitHub repo to a hosted
site.

Additional workflows intended for the codespaces demo have been removed.